[diguagege]：完善了测试模块，需要push Suject.txt和Subscribe.txt到sdcard，要不会崩溃，而且还需要按照Suject: id,title,type Subscribe:id, title, time来录入，逗号改为空格隔开 变更类别：新增逻辑

+++ app/src/main/java/subscribe/diguagege/com/subscribe/base/SubscribeProviders.java
@@ -155,7 +155,7 @@ public class SubscribeProviders extends SQLiteContentProvider {
             case SUBSCRIBE_ID:
                 id = mHelper.insertSubscribe(values);
                 long subjectId = ContentUris.parseId(uri);
-                int subscribeId = (int) values.get(SubscribeContract.Subscribe.EVENT_ID);
+                long subscribeId = (long) values.get(SubscribeContract.Subscribe.EVENT_ID);
                 if (subscribeId >= 0) {
                     mHelper.insertLinked(subjectId, subscribeId);
                     sendUpdateNotification(subscribeId);
